Navigate menus with the same keys as the map.
